s/log: do not roll when there are no segments open
Truncating the log when there are no opened log segments caused redpanda
to crash. It was caused by executing action on empty optional.
Preventing this from happening by skipping rolling the logs after
truncation of empty log as there are no log segments to roll.

Message-Id: <20191214160348.44130-1-michal@vectorized.io>

+++ src/v/storage/log.cc
@@ -149,7 +149,10 @@ future<> log::do_truncate(model::offset o, model::term_id term) {
     //  3.
     auto erased = _segs.remove(std::move(names_to_delete));
     auto f = make_ready_future<>();
-
+    // do not roll when we do not have segment opened
+    if (!_appender) {
+        return f;
+    }
     // 4.
     f = parallel_for_each(erased, [](segment_reader_ptr i) {
         return i->close().then([i] { return remove_file(i->get_filename()); });
